backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
websockets>=12.0
minio>=7.2.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
